+++ retrocache/src/test/java/com/andiag/retrocache/cache/ToStringConverterFactory.java
@@ -1,9 +1,8 @@
-package com.andiag.retrocache.utils;
+package com.andiag.retrocache.cache;
 
 import java.io.IOException;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
-import java.nio.charset.Charset;
 
 import okhttp3.MediaType;
 import okhttp3.RequestBody;
@@ -18,31 +17,24 @@ import retrofit2.Retrofit;
 public class ToStringConverterFactory extends Converter.Factory {
     private static final MediaType MEDIA_TYPE = MediaType.parse("text/plain");
 
-
     @Override
     public Converter<ResponseBody, ?> responseBodyConverter(Type type, Annotation[] annotations,
                                                             Retrofit retrofit) {
         if (String.class.equals(type)) {
             return new Converter<ResponseBody, String>() {
-                @Override
-                public String convert(ResponseBody value) throws IOException {
-                    return new String(value.bytes(), Charset.defaultCharset());
+                @Override public String convert(ResponseBody value) throws IOException {
+                    return value.string();
                 }
             };
         }
         return null;
     }
 
-    @Override
-    public Converter<?, RequestBody> requestBodyConverter(Type type,
-                                                          Annotation[] parameterAnnotations,
-                                                          Annotation[] methodAnnotations,
-                                                          Retrofit retrofit) {
-
+    @Override public Converter<?, RequestBody> requestBodyConverter(Type type,
+                                                                    Annotation[] parameterAnnotations, Annotation[] methodAnnotations, Retrofit retrofit) {
         if (String.class.equals(type)) {
             return new Converter<String, RequestBody>() {
-                @Override
-                public RequestBody convert(String value) throws IOException {
+                @Override public RequestBody convert(String value) throws IOException {
                     return RequestBody.create(MEDIA_TYPE, value);
                 }
             };